coaching/pyproject.toml
[project]
name = "purposepath-coaching"
version = "1.0.0"
description = "PurposePath Coaching Service - AI-powered coaching conversations"
authors = [
    {name = "PurposePath Team", email = "dev@purposepath.ai"}
]
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
keywords = ["aws", "lambda", "ai", "coaching", "bedrock", "serverless", "purposepath"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    # Core FastAPI dependencies
    "fastapi>=0.109.0,<0.115.0",
    "mangum>=0.17.0,<0.18.0",
    "uvicorn[standard]>=0.27.0,<0.32.0",
    "pydantic>=2.5.0,<3.0.0",
    "pydantic-settings>=2.1.0,<3.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    # AWS SDK and extensions
    "boto3>=1.34.0,<2.0.0",
    "botocore>=1.34.0,<2.0.0",
    "boto3-stubs[dynamodb,s3,secretsmanager,bedrock,ses]>=1.34.0",
    # LLM and AI - LangChain ecosystem with multi-provider support
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-aws>=0.2.12",  # ChatBedrockConverse support
    "langchain-anthropic>=0.2.0",
    "langchain-openai>=0.2.0",
    "openai[aiohttp]>=1.60.0",  # Faster async HTTP client
    "langgraph>=0.2.0",
    "langsmith>=0.1.0",
    "tiktoken>=0.8.0",
    # Database and caching
    "redis>=5.0.0,<6.0.0",
    "redis[hiredis]>=5.0.0,<6.0.0",
    # Utilities and security
    "python-jose[cryptography]>=3.3.0,<4.0.0",
    "python-multipart>=0.0.6,<1.0.0",
    "httpx>=0.25.0,<1.0.0",
    "structlog>=23.2.0,<25.0.0",
    "pyyaml>=6.0.1,<7.0.0",
    # Data validation
    "email-validator>=2.0.0,<3.0.0",
    "jinja2>=3.1.6",
    "html2text>=2024.2.26,<2025.0.0",
    "beautifulsoup4>=4.14.3",
    
    # Google Gen AI SDK (replaces deprecated google-cloud-aiplatform for generative AI)
    # Migration: https://cloud.google.com/vertex-ai/generative-ai/docs/deprecations/genai-vertexai-sdk
    "google-genai>=1.56.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=7.4.0,<9.0.0",
    "pytest-asyncio>=0.21.0,<1.0.0",
    "pytest-mock>=3.11.1,<4.0.0",
    "pytest-cov>=4.1.0,<6.0.0",
    "pytest-xdist>=3.3.1,<4.0.0",
    
    # Code quality
    "black>=23.7.0,<25.0.0",
    "ruff>=0.1.0,<1.0.0",
    "mypy>=1.4.1,<2.0.0",
    "pre-commit>=3.3.3,<4.0.0",
    
    # Development tools
    "httpx>=0.25.0,<1.0.0",
    "faker>=20.0.0,<25.0.0",
    "jupyter>=1.0.0,<2.0.0",
    "ipython>=8.12.0,<9.0.0",
    "types-PyYAML>=6.0.12.20240808",
    "types-python-jose>=3.3.4.20240106",
    "types-requests>=2.32.0.20241016",
]

build = [
    "build>=0.10.0,<2.0.0",
    "wheel>=0.40.0,<1.0.0",
]

[project.urls]
Homepage = "https://github.com/purposepath/api"
Documentation = "https://docs.purposepath.ai"
Repository = "https://github.com/purposepath/api.git"
"Bug Tracker" = "https://github.com/purposepath/api/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.11.1",
    "black>=23.7.0",
    "ruff>=0.1.0",
    "mypy>=1.4.1",
    "ipython>=8.12.0",
]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "ARG", "SIM", "RUF"]
ignore = [
    "E501",   # Line too long
    "B008",   # FastAPI Depends() in defaults is standard pattern
    "N999",   # Invalid module name (triggered by repo path in CI)
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"**/tests/**/*" = ["B011", "ARG001", "ARG002", "S101", "B007", "B017", "SIM118", "RUF043"]

[tool.mypy]
python_version = "3.11"
# Use strict settings but allow untyped decorators for FastAPI
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_defs = true
no_implicit_optional = true
no_implicit_reexport = true
strict_equality = true
warn_redundant_casts = true
warn_return_any = true
warn_unused_configs = true
warn_unused_ignores = true
# FastAPI compatibility - allow untyped decorators
disallow_untyped_decorators = false
ignore_missing_imports = true

# Note: FastAPI decorator warnings are ACCEPTED as framework limitation
# These "Untyped decorator makes function X untyped [misc]" warnings do not
# indicate actual type safety issues. FastAPI lacks complete type stubs for
# decorators, and types-fastapi does not exist. This is industry-standard.
# CI configured with continue-on-error: true for MyPy checks.
# See: Issue #68 - All 148 concrete type errors resolved (70.8% completion)

# Per-module configuration for FastAPI routes
[[tool.mypy.overrides]]
module = [
    "coaching.src.api.routes.*",
    "coaching.src.api.main",
]
# Allow untyped decorators in FastAPI route files
disallow_untyped_decorators = false
warn_return_any = false

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = [".", "src"]
asyncio_mode = "auto"
markers = [
    "unit: marks tests as unit tests (fast, isolated)",
    "integration: marks tests as integration tests (slower, external dependencies)",
    "e2e: marks tests as end-to-end tests",
]
